+++ src/Loader/FileLoader.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace AutoMapper\Loader;
 
+use AutoMapper\Exception\RuntimeException;
 use AutoMapper\Generator\MapperGenerator;
 use AutoMapper\Metadata\MapperMetadata;
 use AutoMapper\Metadata\MetadataFactory;
@@ -128,7 +129,7 @@ final class FileLoader implements ClassLoaderInterface
         $fp = fopen($file, 'w');
 
         if (false === $fp) {
-            throw new \RuntimeException(sprintf('Could not open file "%s"', $file));
+            throw new RuntimeException(sprintf('Could not open file "%s"', $file));
         }
 
         if (flock($fp, LOCK_EX)) {