+++ src/Provider/ProviderRegistry.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace AutoMapper\Provider;
 
+use AutoMapper\Exception\InvalidArgumentException;
+
 /**
  * @internal
  */
@@ -34,7 +36,7 @@ final readonly class ProviderRegistry
     public function getProvider(string $id): ProviderInterface
     {
         if (!\array_key_exists($id, $this->providers)) {
-            throw new \InvalidArgumentException(sprintf('Provider with id "%s" not found.', $id));
+            throw new InvalidArgumentException(sprintf('Provider with id "%s" not found.', $id));
         }
 
         return $this->providers[$id];